requirements.txt
annotated-types==0.7.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudscraper==1.2.71
Deprecated==1.2.14
Flask==3.0.3
Flask-Limiter==3.7.0
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.12.0
loguru==0.7.2
markdown-it-py==2.2.0
MarkupSafe==2.1.5
mdurl==0.1.2
nest-asyncio==1.6.0
ordered-set==4.1.0
packaging==24.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pyparsing==3.1.2
python-dotenv==1.0.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.3.4
rsa==4.9
tqdm==4.66.4
typing_extensions==4.12.0
uritemplate==4.1.1
urllib3==2.2.1
websocket-client==1.8.0
Werkzeug==3.0.3
wrapt==1.16.0
